asterisk/services/requirements.txt
ari==0.1.3
gevent==1.1.2
greenlet==0.4.10
OdooRPC==0.4.3
Pillow==3.3.1
requests==2.11.1
six==1.10.0
swaggerpy==0.2.1
websocket-client==0.37.0
wheel==0.24.0
